refactor: update Sort references to use the new request module and remove obsolete code
