+++ notionrs/src/object/request/mod.rs
@@ -1,2 +1,3 @@
 pub mod filter;
 pub mod search;
+pub mod sort;
